refactor(jarm): move JARM core to Features.ResponseObject namespace
